requirements2.txt
keras==2.4.3
gunicorn==19.0.0
scikit-learn==0.23.2
python-dateutil==2.8.1
app==0.0.1 
python-application==2.8.0
pillow==6.2.0
streamlit==0.73.1
numpy==1.18.1
pandas==1.1.4
tensorflow-cpu==2.3.1
h5py==2.10.0